assistant/requirements.txt
# Core Django
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.10
pgvector==0.2.4

# Background Tasks
celery==5.4.0
redis==5.2.1

# AI Services
openai==1.59.3
tiktoken==0.5.2

# File Processing
PyPDF2==3.0.1
python-magic==0.4.27
Pillow==10.4.0

# File Storage (S3/MinIO)
boto3==1.35.84
django-storages==1.14.4
minio==7.2.11

# Authentication & Security
PyJWT==2.10.1
cryptography==43.0.3
phonenumbers==8.13.51
django-ratelimit==4.1.0
django-bleach==3.1.0
django-csp==4.0

# Google Calendar Integration
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.155.0

# HTTP Requests
requests==2.32.3

# Utilities
python-decouple==3.8
python-dateutil==2.9.0
qrcode[pil]==7.4.2

# Development & Testing
pytest==8.3.4
pytest-django==4.9.0

# Production WSGI/ASGI
gunicorn==23.0.0
uvicorn==0.32.1

# Additional security (optional but recommended)
bleach==6.2.0

# Monitoring and logging (optional)
sentry-sdk==2.19.2
